pyproject.toml
[tool.poetry]
name = "kr-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
coreapi = "^2.3.3"
setuptools = "^75.1.0"
django-cors-headers = "^4.4.0"
drf-spectacular = "^0.27.2"
requests = "^2.32.3"
celery = "^5.4.0"
eventlet = "^0.37.0"
redis = "^5.1.0"
django-celery-beat = "^2.7.0"
python-telegram-bot = "^21.6"
pytelegrambotapi = "^4.23.0"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
